Create README.md
Is it true that our information on the internet, photos, searches, miles traveled, timestamps, etc. are assests. Do i myself have money, underlying or unrealized investments/cash that im not aware of? Yes, I believe so, but mine are not linked chains because of lost devices, accounts and passwords. If is not actual tender that is fine because my memories are truley priceless and I want them all back. Help me to link my memories back together again.